Made separate components/ updated routes and names
