+++ src/components/Dashboard/DashboardContainer.js
@@ -1,8 +1,8 @@
 import React from "react";
-import { Route } from "react-router-dom";
+import { Route, Switch } from "react-router-dom";
 import Dashboard from "./components/Dashboard";
 import SideBar from "./components/SideBar";
-import SearchBar from "./components/SearchBar";
+import FoodSearch from "./components/FoodSearch/FoodSearch";
 // import ProtectedRoute from "../utils/ProtectedRoute";
 
 const DashboardContainer = () => {
@@ -11,7 +11,7 @@ const DashboardContainer = () => {
       <SideBar />
       <div className="dashboard mt-4 wrapper flex-column">
         <Route exact path="/dashboard" component={Dashboard} />
-        <SearchBar />
+        <Route path="/dashboard/food-search" render={() => <FoodSearch />} />
       </div>
     </div>
   );
