+++ src/components/Dashboard/styles/_dashboard.scss
@@ -3,3 +3,14 @@
   width: 75%;
   left: 20%;
 }
+
+.foodLog{
+  border-radius: 25px;
+  width: 90%;
+  margin-bottom: 5%;
+  position: relative;
+  padding: 2%;
+  left: 5%;
+  background-color: #dedfdf17;
+  box-shadow: -15px -15px 15px rgba(255,255,255,0.5), 15px 15px 18px rgba(59, 59, 59, 0.08);
+}
