+++ src/components/Home/components/SignIn.js
@@ -4,7 +4,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { Link } from "react-router-dom";
 import { Redirect } from "react-router-dom";
 import { ReactComponent as SignInImage } from "../../../assets/GirlComptr.svg";
-import { authenticate } from "../../../state/Slices/slices";
+import { authenticate } from "../../../state/slices/slices";
 import Footer from "./LandingPage/components/Footer";
 
 const SignIn = () => {
@@ -31,7 +31,7 @@ const SignIn = () => {
     setUser({ ...user, [e.target.name]: e.target.value });
   };
 
-  if (isAuthenticated) return <Redirect to="/dashboard" />;
+  if (isAuthenticated) return <Redirect to="/protected" />;
 
   return (
     <div>
